virst version witout output and custom strcmp

+++ str_sort.h
@@ -1,6 +1,8 @@
 #ifndef STR_SORT_H
 #define STR_SORT_H
 
-void str_sort(char **line_ptrs, int str_num);
+#include <stdlib.h>
+
+void str_sort(char **line_ptrs, size_t str_num);
 
 #endif
